Added panels and game class. Worked on cast rays function, but it doesnt work

+++ RaycasterGame/Map.cpp
@@ -51,7 +51,7 @@ void Map::draw(int windowWidth, int windowHeight)
 	}
 }
 
-bool Map::validPosition(int xPos, int yPos)
+bool Map::validPosition(float xPos, float yPos)
 {
 	int x0;
 	int y0;
